AUEffectProcessor: Factor out AudioUnitManager
